Add filtering (#2014)

+++ test/Performance/MSTest.Performance.Runner/Steps/PerfviewRunner.cs
@@ -7,22 +7,24 @@ using System.IO.Compression;
 using System.Runtime.InteropServices;
 using System.Text;
 
-using MSTest.Performance.Runner.Runners;
-
-namespace MSTest.Performance.Runner.Runner;
+namespace MSTest.Performance.Runner.Steps;
 
 internal class PerfviewRunner : IStep<BuildArtifact, Files>
 {
     private const string PrefViewDownload = "https://github.com/microsoft/perfview/releases/download/v3.1.7/PerfView.exe";
     private readonly string _argument;
     private readonly string _reportFileName;
+    private readonly bool _includeScenario;
+    private readonly CompressionLevel _compressionLevel;
 
     public string Description => "run under PerfView";
 
-    public PerfviewRunner(string argument, string reportFileName)
+    public PerfviewRunner(string argument, string reportFileName, bool includeScenario = false, CompressionLevel compressionLevel = CompressionLevel.Fastest)
     {
         _argument = argument;
         _reportFileName = reportFileName;
+        _includeScenario = includeScenario;
+        _compressionLevel = compressionLevel;
     }
 
     public async Task<Files> ExecuteAsync(BuildArtifact payload, IContext context)
@@ -35,8 +37,8 @@ internal class PerfviewRunner : IStep<BuildArtifact, Files>
 
         string perfViewExecutable = await PerfviewExecutable();
         StringBuilder commandLine = new();
-        commandLine.Append(CultureInfo.InvariantCulture, $"\"/DataFile:{Path.Combine(Path.GetDirectoryName(payload.TestHost.FullName)!, "DataFile.etl")}\" /NoGui {_argument} ");
-        commandLine.Append(CultureInfo.InvariantCulture, $"run {payload.TestHost.FullName} ");
+        commandLine.Append(CultureInfo.InvariantCulture, $" \"/DataFile:{Path.Combine(Path.GetDirectoryName(payload.TestHost.FullName)!, "DataFile.etl")}\" /AcceptEULA /NoGui {_argument} ");
+        commandLine.Append(CultureInfo.InvariantCulture, $"run \"{payload.TestHost.FullName}\" ");
 
         ProcessStartInfo processStartInfo =
         new(await PerfviewExecutable(), commandLine.ToString())
@@ -49,6 +51,7 @@ internal class PerfviewRunner : IStep<BuildArtifact, Files>
 
         Console.WriteLine($"Perfview command: '{processStartInfo.FileName} {processStartInfo.Arguments}'");
 
+        bool succeded = false;
         ManualResetEvent killTheProcess = new(false);
         Process process = Process.Start(processStartInfo)!;
         process.EnableRaisingEvents = true;
@@ -56,9 +59,13 @@ internal class PerfviewRunner : IStep<BuildArtifact, Files>
         process.OutputDataReceived += (sender, args) =>
         {
             Console.WriteLine(args.Data);
-            if (args.Data == "Press enter to close window.")
+            if (args?.Data?.Contains("SUCCESS: PerfView") == true)
             {
-                process.StandardInput.WriteLine();
+                succeded = true;
+            }
+
+            if (args?.Data == "Press enter to close window.")
+            {
                 killTheProcess.Set();
             }
         };
@@ -71,10 +78,31 @@ internal class PerfviewRunner : IStep<BuildArtifact, Files>
 
         process.WaitForExit();
 
+        if (!succeded)
+        {
+            throw new InvalidOperationException("Perview command didn't succeed.");
+        }
+
         string sample = Path.Combine(Path.GetTempPath(), _reportFileName);
         File.Delete(sample);
+
         Console.WriteLine($"Compressing to '{sample}'");
-        ZipFile.CreateFromDirectory(payload.TestAsset.TargetAssetPath, sample, CompressionLevel.SmallestSize, includeBaseDirectory: true);
+        if (_includeScenario)
+        {
+            ZipFile.CreateFromDirectory(payload.TestAsset.TargetAssetPath, sample, _compressionLevel, includeBaseDirectory: true);
+        }
+        else
+        {
+            string reportDirectory = Path.GetDirectoryName(payload.TestHost.FullName)!;
+            string dataFileDirectory = Path.Combine(reportDirectory, "DataFile");
+            Directory.CreateDirectory(dataFileDirectory);
+            foreach (var item in Directory.GetFiles(reportDirectory, "DataFile.*"))
+            {
+                File.Move(item, Path.Combine(dataFileDirectory, Path.GetFileName(item)!));
+            }
+
+            ZipFile.CreateFromDirectory(dataFileDirectory, sample, _compressionLevel, includeBaseDirectory: true);
+        }
 
         return new Files(new[] { sample });
     }
